Update the version and use as image version (#107)
Update the OTEAPI Services version using the invoke task `setver`.
The version is made up of
`<API version>.<CalVer (YYYY0M0D)>.<Build #>`.

Keep using the "latest" tag version as well, pushing both.

Whenever a push to `master` happens, the `oteapi` container image will
be updated with a "latest" version tag.
It will also be updated with a version tag that matches the current
OTEAPI Services version.

The version will be updated and an amend commit will be done,
reusing the latest commit message.
This will then be force-pushed to `master` and the dependabot
branch will be updated as well as a docker image built and deployed.

Also:
Clean up README a bit.
Only allow `setver` invoke task if in CI/CD environment.

Ensure only a single run is ever in progress:
This is for the CI/CD workflow that updates the `master` branch - as
well as publishes a Docker image and updates the `ci/dependabot-updates`
branch.