Granularize python dependencies
Split them into "std", "editable", and "external".
"std" are the "normal" packages that show up during `pip freeze`, e.g.,
in the form of `oteapi-core==0.1.6`.
"editable" are all the ones showing up in `pip list`, but not `pip
freeze --exclude-editable`.
"external" is everything in `pip freeze` that isn't in the "std" format.