Cache only the set parameters in the configurations (#405)
Also, avoid setting the `token` parameter if there is no reason to.
Also, remove the `--wait-for-client` option when running the service via
`debugpy`. This will avoid running the server as long as VS Code has not
started a session. If this is specifically necessary, e.g., when
debugging startup events, then this can be added in locally.